Move helpers to their corresponding components
