+++ src/types/Entry.ts
